+++ Test.Integration.Net/EventsAPI/LoginLogout.cs
@@ -1,9 +1,9 @@
 using Sift;
 using System.Collections.ObjectModel;
-using Test.Integration.Net7.Uitlities;
+using Test.Integration.Net.Uitlities;
 using Xunit;
 
-namespace Test.Integration.Net7.EventsAPI
+namespace Test.Integration.Net.EventsAPI
 {
     public class LoginLogout
     {
