+++ Test.Integration.Net/EventsAPI/Notifications.cs
@@ -1,51 +1,47 @@
 using Sift;
-using Test.Integration.Net7.Uitlities;
+using Test.Integration.Net.Uitlities;
 using Xunit;
 
-namespace Test.Integration.Net7.EventsAPI
+namespace Test.Integration.Net.EventsAPI
 {
-    public class Verifications
+    public class Notifications
     {
         private readonly EnvironmentVariable environmentVariable = new();
         private readonly string ApiKey;
-        private readonly string SessionId;
         private readonly string UserId;
-        private readonly string VerifiedValue;
-        public Verifications()
+        private readonly string SessionId;
+        private readonly string NotifiedValue;
+        public Notifications()
         {
             ApiKey = environmentVariable.ApiKey;
-            SessionId = environmentVariable.session_id;
             UserId = environmentVariable.user_id;
-            VerifiedValue = environmentVariable.verified_value;
+            SessionId = environmentVariable.session_id;
+            NotifiedValue = environmentVariable.notified_value;
         }
         [Fact]
-        public void VerificationTest()
+        public void SecurityNotificationTest()
         {
             var sift = new Client(ApiKey);
-            var sessionId = "sessionId";
-            var verification = new Verification
+            var securityNotification = new SecurityNotification
             {
                 user_id = UserId,
                 session_id = SessionId,
-                status = "$pending",
+                notification_type = "$email",
+                notified_value = NotifiedValue,
+                notification_status = "$sent",
                 browser = new Browser
                 {
                     user_agent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_12_3) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/56.0.2924.87 Safari/537.36",
                     accept_language = "en-US",
                     content_language = "en-GB"
                 },
-                verified_event = "$login",
-                verified_entity_id = sessionId,
-                verification_type = "$sms",
-                verified_value = VerifiedValue,
-                reason = "$user_setting",
                 brand_name = "sift",
                 site_domain = "sift.com",
                 site_country = "US"
             };
             EventRequest eventRequest = new EventRequest()
             {
-                Event = verification
+                Event = securityNotification
             };
             EventResponse res = sift.SendAsync(eventRequest).Result;
             Assert.Equal("0", res.Status.ToString());
