+++ Test.Integration.Net/EventsAPI/Sessions.cs
@@ -1,8 +1,8 @@
 using Sift;
-using Test.Integration.Net7.Uitlities;
+using Test.Integration.Net.Uitlities;
 using Xunit;
 
-namespace Test.Integration.Net7.EventsAPI
+namespace Test.Integration.Net.EventsAPI
 {
     public class Sessions
     {
