+++ Test.Integration.Net/EventsAPI/Transactions.cs
@@ -2,10 +2,10 @@ using Sift;
 using System;
 using System.Collections.ObjectModel;
 using System.Diagnostics;
-using Test.Integration.Net7.Uitlities;
+using Test.Integration.Net.Uitlities;
 using Xunit;
 
-namespace Test.Integration.Net7.EventsAPI
+namespace Test.Integration.Net.EventsAPI
 {
     public class Transactions
     {
