+++ Test.Integration.Net/local.settings.json
